WL#17062 Add option for guided load to ALTER TABLE SECONDARY_LOAD (parser changes)
An optional clause is added to ALTER TABLE ... SECONDARY_LOAD
so that the user can control whether guided load should be
used or not.

The new syntax is:
  ALTER TABLE table_name SECONDARY_LOAD
    [PARTITION (...)] [GUIDED {ON | OFF}]

If guided load is supported, the default behavior is to use it.
That is, when a GUIDED clause is not specified, guided load
will happen.

Change-Id: I4edcc507368e426b0d8009788e3a7f6a328ca671