Bug#35451459: MySQL Server crashes when executing query
unpack_partition_info() is called before allocating bitmaps,
which needs to be done prior to generated columns as
they will call fix_fields and functions might want to access bitmaps

This was a regression from fix of bug#33142135.
It was fixed in 9.3.0 but not in 8.0+ versions.
This fix is a backport of Bug#35044654.

Change-Id: Iedab09f3cfae730e0db3da7cde7af61a4b22b248