Bug#38021487: Assertion `false && "Inconsistent row counts for different AccessPath objects."' failed.
Queries in which some tables could be pruned away due to alway false
join conditions, sometimes raised assertion failures because the
cardinality estimates were inconsistent.

The inconsistency happened because the ClearImpossibleJoinConditions()
step of MakeJoinHypergraph() removed join conditions in the pruned
subtree of the inital acyclic hypergraph, and some of these join
conditions could be added back when a later step in
MakeJoinHypergraph() added cycle edges within the pruned subtree.
Since the selectivity of these join conditions had already been
accounted for the first time they were seen, before
ClearImpossibleJoinConditions(), adding them back a second time caused
double-counting of the selectivity for some join orders, which led to
inconsistencies in the estimates and assertion failures.

Since the subtree of the join is pruned away in the final query plan
anyway, considering more join orders for the tables in that subtree is
a pointless exercise which won't lead to a better plan being found.
The patch skips adding cycle edges in the hypergraph for those parts
that are pruned away, and thereby avoiding the reintroduction of
pruned join conditions and double-counting their selectivity.

Change-Id: I1d272734423dbf3be594c150892b0c851036c710