Bug#38168435 routertest_integration_routing_direct fails sporadically
Running the server_greeeting_error tests of
routertest_integration_routing_direct on a low-core machine sometimes
makes the test fail.

  $ taskset -c 0
  ./runtime_output_directory/routertest_integration_routing_direct
  --gtest_filter=*server_greeting_error*

The test fails with "1040 Too many connections" when the test expects
success.

Background
==========

In the test a connection is closed and a new connection is opened and it
is assumed that this close is immediately seen by the server, but that's
not the case.

Closing the old connection and openening the new connection may be in
different threads leading to the new connection getting denied because
the old connection hasn't cleanup up yet in the server.

Change
======

- retry to connect for max 500ms in case of error 1040

Change-Id: I2325b4324915bbbcdd2a520a2fec789fd1ba903f