Bug#38168569 routertest_component_state_file ip_v6 tests fail on Windows
The test configures the mocks servers to bind to the ::1
address but then still sometimes uses 127.0.0.1 to
connect to them. That does not work on Windows, only
the connections on ::1 are accepted there.

This patch changes the test code to always use ::1 when
connecting to the mock servers configured for ipv6.

Change-Id: Ib25fdf422e66772239bf5aebe5b19ddd538873e4