Bug#38170645 Missing GCOV coverage for "merged" unit tests
For some unknown reason, gcov fails to report coverage of our "merged" unit tests.
The solution is to build an executable for each individual test file,
and to do ADD_TEST for it.
We also add EXCLUDE_FROM_GCOV for the "merged" unit tests,
since everything is covered by the individual tests.

Change-Id: I456d779f0e326ecce37905f106a8d6284ca0f09d