Bug#38174467 routertest_component_routing_splitting fails on Windows
The test sets up the hostname of the mock_server nodes in the
metadata as "localhost". The mock_servers are configured to
bind only to "127.0.0.1" interface. As a result on Windows,
the Router can't connect to the metadata servers (mocks) using
the host:port fetched from the metadata.

This patch changes the hostname set in the metadata to "127.0.0.1".
It also adjusts the test expectations following this change.

Change-Id: Idde646290930b7dad010257eb76974b2069810d8