BUG#37977799 Memory leak issue with router scripting tests
The execution of javscript files is leaking 2K per file per
JavaScript context created, the leak comes from Graal itself and
there's no way to resolve it, this patch is to minimize it's impact.

The context pool would create as many contexts as needed to satisfy
the client load, a flaw in the logic was identified which caused
the contexts to be stuck in a release step before moving them to the
pool, this lead to a continuous creation of more contexts from time
to time, increasing the impact of the leak.

The release logic was refactored in a way that contexts that are
are queued for releasing only when needed, otherwise they are moved
to the pool right away, this reduced dramatically the amount of
created contexts (stress testing showed that only N contexts were
created, where N is the number of io threads in the router), the pool
still has the capacity to create more if needed, but not seen in the
tests. This basically reduced the impact of the leak to the minimal
possible.

Change-Id: I6e6a3a082ff58b6cf4e445f8cbd5892f4b3f874e