Bug#38181929 Router may not stop accepting ports if the md refresh fails on GR state update
If the Router updates the state of the GR iterating over
its node and can't find any usable nodes it would proceed
without error in the code (only in the logfile) clearing
the nodes list. Upper layer seen the empty list and no
error would assume there was no change, so there is no need
to close the acceptor ports.

This patch fixes the code so that when this error condition
is discoved it is propagated to the upper layers. Now
the calling code is closing the acceptor ports as expected.

Change-Id: I77360cf1d06fa58946287f4888d1424a23eeab7c