Bug#38168051: Support for GROUPING() function with               primitive GROUP BY
GROUPING function was not allowed in a query without
the presence of group by modifiers like ROLLUP.
We lift that limitation now. Although, using GROUPING
function without group by modifiers might not be
useful, standard allows it.
Since the presence of GROUPING function was tightly
coupled with the presence of ROLLUP in the query,
some of the semantic checks like arguments to a
GROUPING function to be present in a GROUP by clause
were done when setting up ROLLUP in resolve_rollup().
It's now changed to be independent of ROLLUP.
These checks cannot be carried out while resolving
the GROUPING function as GROUP BY is still not
resolved at the time.

Change-Id: Icfbcf349c89998a3712ee7e4b7e7cff850d3a91d