Bug#38193362 Switch to OpenSSL 3 on Solaris
We are still building with OpenSSL 1.0.2 as "system" OpenSSL on
Solaris. OpenSSL 3 has been available for quite some time.
As of SRU81, OpenSSL 1.0.2 is gone, so switch to OpenSSL 3.0

We also copy /usr/openssl/3/bin/openssl to bin/my_openssl,
so that mtr will use the correct version during testing.

Cosmetics: add a few more lines to RESET_SSL_VARIABLES, and sort it.

Change-Id: I2e82a15c9a7782ec7707d8d22eb8fb1602ceb43d