Bug#38132603 Router unable to get metadata when Secondaries are lagging and do not have router_id yet
In the following scenario:
- the target cluster is part of a ClusterSet
- the first metadata server from the Router list in `state.json` file is a SECONDARY node
- that server is lagging behind the PRIMARY and still did not replicate the metadata
  after the bootstrap when the Router was launched
... the Router fails to open accepting points when started.

The problem is the discrepancy between the Router local configuration (config file and
dynamic state file) and what it sees in the metadata. The Router was configured to
work with a ClusterSet but when it wants to check the dynamic setting for that in
the metadata in `routers.attributes->$.bootstrapTargetType` there is no entry
in routers table (not replicated yet). This leads to the scenario where it assumes
the default: treat a target as a standalone Cluster. That is not consistent with
the configuration created during the bootstrap and as a result the Router gets
stuck and never opens accepting ports.

This patch fixes handling of such scenario. Now whenever the metadata server does
not see an entry for a given router id in the metadata, this metadata server
is discarded and the Router proceedes to the next metadata server from the list
for a valid metadata.

Change-Id: I07c2c69d111b98dd84004c07d4ae2cbc7813228e