Bug#37833432 Content set: directoryIndexDirective order violated
Issue
=====

directoryIndexDirective should be an ordered list, but requesting
a resource that has directoryIndexDirective set showed that this order
is not taken into account. This is caused by the fact that we have no
control over the order in which Content Files are created - therefore
it is not possible to set a directoryIndexDirective handler from
Content File level.

Fix
===

Creating directoryIndexDirective handler is relegated to the parent.

Change-Id: I49e6392f35d71fcfdd8b2f0d6c63e2895eb72345