Bug#38211194: Hypergraph does not push non-equality predicates to existing cycle edges
The hypergraph optimizer only attempts to create cycles in the
hypergraph using simple equality predicates. When there is a
non-equality join predicates between two tables that are connected by
a cycle edge created from such a simple equality predicate, the
non-equality predicate is however not put on that simple edge.
Instead, a wider hyperedge is constructed, and the predicate is pushed
onto that hyperedge.

The wider hyperedge does not help finding more join orders than those
that can be found by the simple edges. It only adds redundancy to the
hypergraph, and makes the planning more complicated. It would be
better if it had pushed the non-equality predicate to the simple edge
between the two tables instead, and it didn't create the wider edge.

Example:

  SELECT 1 FROM t1, t2, t3
  WHERE t1.x = t2.x AND t1.y = t3.y AND
        t2.z = t3.z AND t2.w <> t3.w

When creating a hypergraph from this query, it first adds a simple
edge between t1 and t2 using the t1.x = t2.x predicate. Then it adds a
simple edge between t1 and t3 using the t1.y = t3.y predicate. The
hypergraph is now complete in the sense that all the tables are
connected. When it gets to the t2.z = t3.z predicate, it sees that
this predicate could be used for creating a cycle in the hypergraph,
and stores it in cycle_inducing_edges for later. Finally, when it
inspects the t2.w <> t3.w predicate, is determines that it is not a
candidate for a cycle edge, as it is not an equality predicate. Since
it cannot be put on the simple t1-t2 or t1-t3 edges created so far,
the t1-t3 edge is widened to {t1,t2}-t3, and the predicate is put
there.

This patch changes how the t2.w <> t3.w predicate is handled. Before
giving up and widening an edge, it checks if we have already found a
predicate which will create a cycle edge between the same two tables,
and, if so, put the predicate on that edge instead.

Change-Id: I12f9c0d3405fa86bd64b59d0f10ea245f43b24cd