Bug#37092159:  Client Server Handshake error while authenticating    authentication_oci plugin with --default-auth paramater.
When the server and the client's default authentication methods are
different at connect time, the client just disposes of the cached
challenge by the server and calls the client authentication method
to conclude the first exchange before attempting a 2nd one with the
actual authentication method for the user. As a result the client
plugin may not have a challenge to read at all. And it should
error out in this case.

But authentication_oci was printing a message to stderr before
erroring out.
Removed the message and added some more DBUG support for the plugin.

Change-Id: Ia17911e705896013d60f517336ade6eaf3d1c746