Bug#38151265 Assertion Failure in Backup::execDEFINE_BACKUP_REQ 2/N
Problem:
During a backup operation the master node receives a
BACKUP_FRAGMENT_CONF - meaning that a fragment was
successfully  and backed up, for a fragment that master already
has identified as scanned. This causes a crash in the master
because the code includes a require statement designed to
prevent a fragment from being marked as backed up more than once

The root cause of this issue lies in a bug within the master
node's handling of the BACKUP_FRAGMENT_REF signal. When the
master receives a BACKUP_FRAGMENT_CONF signal, it needs to
iterate through all tables and fragments to locate the
corresponding fragment and mark it as scanned, among other
operations. However, due to the identified bug, the fragment
that gets marked as scanned can be incorrect, leading to the
master node erroneously believing a fragment has been
confirmed twice.

This path adds a new NDBAPI test to testBackup reproducing
the bug scenario.

Change-Id: I14b3eea3e5ac3fda5cdacfb9eafe02093ed9606c