Bug#38151265 Assertion Failure in Backup::execDEFINE_BACKUP_REQ 3/N
Problem:
During a backup operation the master node receives a
BACKUP_FRAGMENT_CONF - meaning that a fragment was
successfully  and backed up, for a fragment that master already
has identified as scanned. This causes a crash in the master
because the code includes a require statement designed to
prevent a fragment from being marked as backed up more than once

The root cause of this issue lies in a bug within the master
node's handling of the BACKUP_FRAGMENT_REF signal. When the
master receives a BACKUP_FRAGMENT_CONF signal, it needs to
iterate through all tables and fragments to locate the
corresponding fragment and mark it as scanned, among other
operations. However, due to the identified bug, the fragment
that gets marked as scanned can be incorrect, leading to the
master node erroneously believing a fragment has been
confirmed twice.

Solution:
Add the tableId and FragmentNo to BACKUP_FRAGMENT_REF.
With these additional identifiers, when the master node receives
a BACKUP_FRAGMENT_REF signal, it can directly access the
corresponding fragmentPtr without the need to iterate over all
tables and fragments. This direct access ensures that the
fragment subsequently marked as scanned is the correct one.

Change-Id: I0e7590a35bf02664379e9ec7d9919bfc358a4b26