Bug#38245274 router.content_files_static_content test fails on UBSAN
In some of the code paths, when the object of entry::ContentFile is
created. it is not  getting fully initialized. When it is marked as
deleted, the content of some of the remaining fields is no relevant.
This object can still get passed to a method like
`log_debug_db_entry_impl` that prints the value of this field, which
makes the UBSAN unhappy as this field value is invalid.

This patch makes sure all the fields of entry::ContentFile are
initialized to valid values.

Change-Id: Id92529a5676577092966e6bd2066a39efbb6ce64