Bug#38175557 - Json Duality View: derived_merge=off optimizer hint not working
When derived_merge=off, the optimizer doest not attempt to merge
derived tables. However, for SELECTs on a JSON duality view,
this hint is not working. Optimizer attempts merge derived table
always irrespective of derived_merge hint state for JSON duality
views.

For DML operations on JSON duality view, currently merging derived
tables is not supported. The checks implemented to prevent this
for DMLs are applicable to SELECT statements, causing the optimizer
to skip merging derived tables even for SELECTs.

To resolve this, materializing derived tables is now permitted for
SELECT statements on JSON duality views and remains restricted only
for DML operations.

Change-Id: Ic46768834495dcaa06778ea3f2ff1440e885cbc0