Bug#38126907 PB2: ndb.ndb_restore_schema_blobs fails due to warning of SHA1 deprecation in output
WL#16955 Deprecate SQL functions MD5() and SHA1() in MySQL Server caused
usage of the SHA1 and MD5 functions to generate deprecation warnings.

Testcase ndb_restore_schema_blobs started failing intermittently despite
being re-recorded to include the warnings as part of the WL.

This is because :
 - Some of the statements using SHA1 in the testcase occur
   just after a table is restored using ndb_restore
 - The MySQL Server is not yet aware that the tables exist
 - Statement execution requires table discovery
 - Table discovery occurs after SQL parsing during 'table
   open'
 - Table discovery clears the thread's 'diagnostics area'
   - Table discovery clears parse-time warnings

In general, this resulted in the SHA1 warnings for the statements
immediately after ndb_restore being lost, and the testcase was
recorded with this output.

However occasionally, the table would already be discovered,
and the warnings would not be discarded, causing a result
mismatch.

-

This patch uses a non warning generating statement
(select count(1)) to force table discovery prior to
the test-functional statements.

This results in the test reliably generating the warnings
as might be expected.

Testcase ndb_mt_backup_ldms had the same problem and is
similarly modified.

Change-Id: I344be07735e0f50f4f0607c13cf0eb1414c285b1