Bug#37382026 MRS bootstrap does not do a cleanup when fails
mysqlrouter_bootstrap utility can be used to bootstrap against InnoDB Cluster,
standalone server with MRS support or InnoDB Cluster with MRS support.
In the third case first the InnoDB Cluster specific bootstrap is done,
after that MRS bootstrap. If the second step fails, no cleanup is done
to revert the InnoDB Cluster bootstrap: there are autogenerated users on
the server and the files/directories created by the bootstrap left behind.

This patch changes the behaviour so that now when the MRS part fails,
the Router attempts to clean up the bootstrap, removing autogenerated
db users and the files/directories create by the bootstrap before the
failure.

Change-Id: Ic46b945a88ae846f8a93da55009793e8ffd08d10