Bug#36684370 crash in Query_block::supported_correlated_scalar_subquery
There is an issue in Item_ref update during subquery to derived
transformation.
Considering below example,

SELECT
    (SELECT f1 FROM (SELECT 1 FROM DUAL) AS dt1
        GROUP BY f2),
    SUM(f1)
    FROM (SELECT * FROM t1) AS dt2;

As part of transforming the scalar subquery, we move the
aggregate SUM(f1) into a derived table, similar to below:

SELECT `derived_1_5`.`sum(c2)` AS `sum(c2)`,
    (SELECT `derived_1_5`.`Name_exp_1`, derived_1_5.Name_exp_2
        FROM (SELECT 3 AS `3`) `v`
        GROUP BY `derived_1_5`.`Name_exp_2`) AS `C`
     FROM (SELECT sum(`r2`.`c2`) AS `sum(c2)`,
                   `r2`.`c3` AS `Name_exp_1`,
                   `r2`.`c2` AS `Name_exp_2`
                 from (`r2`)) `derived_1_5`

As can be seen the original occurrences of f1 and f2 in the
scalar subquery have been replaced with another (outer)
reference: derived_1_5.Name_exp_1 for f1 and
derived_1_5.Name_exp_2 for f2. f2 should be induced as a
hidden field in the select list along with visible
f1 to make it scalar. We lost the hidden flag for
f2 (derived_1_5.Name_exp_2) when the Item_ref is converted
to Item_field during the transformation. During transformation,
view references are replaced with underlying fields
if they are resolved in the current query block, or retained
as view references if their depended_from matches the
transformed block. However, when both conditions are met,
view ref is replaced with the underlying field.
This will cause the hidden flag of the view ref to be lost.
Which lead to the assert as it was no longer a scalar
subquery as expected.

Fix:
The transformation now chooses Item_view_ref when both
conditions are met and constructs a copy of the Item_view_ref
when populating the derived table’s select list, rather than
directly reusing the original object. This preserves the
original’s is_hidden flag, since add_item_to_list() resets
this property when attaching items to the derived table.
By using a cloned reference, hidden status is retained and
restored by replace_view_ref() later. Additional changes
in replace_item_view_ref() includes m_outer_field variable
added to reuse the same Item across multiple reference
to avoid field duplication during temporary table creation.

Change-Id: I3df1878190dd4ad4a2a9af409b1e85e936a16786