Bug#38260769 ndb_restore cannot handle backup with large BackupID(>2147483647).
Backport to 7.6

Valid backup ids are 1..4294967294 inclusive.

ndb_restore is incorrectly treating the backup id as a 32-bit signed
int, so that backup ids > 2147483647 are treated as negative numbers
and their files cannot be opened.

This is fixed, and testing in ndb_restore_misc is enhanced
to cover valid + invalid ids.

Change-Id: Iae4ad4c010a8475c978a5734cb9dcae2a1d931ed