Bug#36326728 Performance degredation using REGEXP in statement              with parameters vs query
Issue:
Queries utilising REGEXP took much longer to execute if they are
run as a prepared statement with parameters rather than direct
queries.

Analysis:
Every time we are checking the match for regular expression for
each row, the regular expression gets re-compiled if the item
is not constant.
Item_param was made const for execution due to which the regular
expression got re-compiled causing the performance degradation.

Fix:
Regular expression need not be compiled for each row for item
that are constant for execution.

Change-Id: I9652ce62ac2afbd4c2bfb910b09c844aef1efc62