BUG#37977799 Memory leak issue with router scripting tests [postfix]
The refactoring done on the JavaScript executor introduced a case
where stalled results caused subsequent requests to give invalid
answers, causing the scripting tests to fail.

This patch introduces a simplified logic that resolves the problem
keeping the advantage of the initial patch which releases the
execution contexts faster, reducing the creation of new executors
and so, reducing the impact of the initial leak.

Change-Id: I522e9f4420ce493e64e25287ce85d5fa767b8926