WL#16658: Hypergraph cost model for skip scan access methods [1/2]
Cost model for index skip scans in hypergraph.

The index skip scan cost calculation function in the range optimiser
is refactored to create a cost-constants calculation function which
can be used by both optimisers. Support is added for cost formula
calculations specific to hypergraph in index skip scan planning.

A JET test is added to check the cost/time correlation for a set of
index skip scan queries using hypergraph on tables which vary in
terms of table size, row length, key length and datatypes.

Some MTR tests are modified to ensure that the queries execute with
skip scans after the cost modifications.

Change-Id: I2f3646ed97b192aaa4693baeeef09a307ee29868
Signed-off-by: Jan Wedvik <jan.wedvik@oracle.com>