WL#16658: Hypergraph cost model for skip scan access methods [2/2]
Cost model for group skip scans in hypergraph.

A cost function specific to hypergraph is added in group skip scan
planning.

A JET test is added to check the cost/time correlation for a set of
group skip scan queries using hypergraph on tables which vary in
terms of table size, row length, key length and datatypes.

Change-Id: I643ec46501037a297ae3c5f8bab03040d1065ed0
Signed-off-by: Jan Wedvik <jan.wedvik@oracle.com>