Bug#37375884 Bootstrapping with --mrs against standalone server accepts --account
mysqlrouter_bootstrap executable allows to bootstrap against the
standalone server that is not part of the InnoDB Cluster.
In that case some of the parameters that are passed to the command
are ignored giving a wrong impression that they may be relevant
for the user in the MRS schema.

This patch changes the behavior so that the bootstrap fails if one
of the following is used to bootstrap against the standalone server:
--account
--account-create
--acount-host

Change-Id: Idc41acdf8c8cafe0ba3310d3aef239f3986dcbbc