BUG#38300798 - mysqld crash in EstimateEqualPredicateSelectivity with prepared statements and materialized views
Background: In case of prepared statements, the derived result
table can be populated with a temporary table, before we make
a decision on whether a storage-engine materialized view
will be used or not. In that case, we need to perform the
cleanup for the TABLE object pointed by derived_result->table.

Problem
=======
cleanup_tmp_tables performs free_tmp_table(tl->derived_result->table);
However, there are still objects in the query block that might refer
to this table object, such as fields.

Solution
========
The freeing of the table object should instead happen at
destroy_tmp_tables, following the same conditions.

Change-Id: I72fbd856c69723b4cbc44658d121f64baa546a81