Bug#37953883 Table_map entry is logged when sql_log_bin = 0
Problem:

When running CREATE/INSERT statements with `sql_log_bin = 0` session
variable on a Ndb Binlog server, then replication would fail due to
Table <X> not existing.

Analysis:

By running some checks through mapping inserted Table_map entries with
the corresponding row changes, then some statements would cause the
checks to fail. The Table_map event is only injected in the Binlog if
the table had any changes to be logged. This is seen from the
any-value carried on the data signal from SUMA.

If some changes in the table result from more than 1 NdbOperation, the
multiple any-values are merged in the NdbApi and can be filtered. In
the Ndb plugin this filtering concludes, from the any-value carried,
that the changes can be: a REPLICA (applied) UPDATE or NOT REPLICA
UPDATE, or NO-LOGGING flag was set (prior to statement) or NOT
NO-LOGGING (i.e. other any-value flag).

Statements that include BLOB (and the like) columns, result in
subsequent NdbOperations to handle the BLOBs (reads/writes/deletes)
and these operations must include the same any-value as the main
NdbOperation of the statement. It was found that these NdbOperation,
created on-the-fly, do not have the any-value set on the Attribute
Info sections of the signal sent to the NDB Kernel.

Solution:

Apply a setAnyValue on every subsequent Blob NdbOperation that I/U/D,
by using the NdbBlob handle as a receiver of the value from the main
NdbOperation at the preExecution phase.

Change-Id: I730dcd84a406898dc28bfea993d91a49745624c4