Bug#38330894 test-unit project is added to the default build on Windows
In the fix for Bug#37833212 "Limit the number of CPUs involved for
cmake targets  'run-profile-suite' and 'test-unit'" we have added an
alias for "test-unit" named "run-unit-tests":

    ADD_CUSTOM_TARGET(run-unit-tests DEPENDS test-unit)

This turns out to have an unitended side-effect of adding test-unit to
the default build when using CMake 4.1.0 and VS 2022 17.12.10.
This then causes CTRL+SHIFT+B (Build Solution) to run all unit tests.

This patch fixes this by explicitily excluding test-unit from the
default build.

Change-Id: I74fe0109c169f076d88a96d699c9eb505b82425c