Bug#38203473: Assertion `is_nullable()' in               Item_typecast_signed::val_int               at sql/item_func.cc [JT_CONST 1/3]
For an empty result, ROLLUP gives a row with grouping
expressions set to NULL and the grand total for
aggregates E.g. COUNT() would give 0. This does not
work as expected when const table optimization is
used by the optimizer.
We now detect that in case of ROLLUP, a row needs
to be generated similar to implicit grouping for a
aggregate functions.

Change-Id: I82052193defd7b38e6cb82ce404dfc26bc59156f