Bug#38203473: Assertion `is_nullable()' in               Item_typecast_signed::val_int               at sql/item_func.cc [ROLLUP 2/3]
For an empty result, ROLLUP generates a row. All grouping
expressions are set nullable while resolving ROLLUP. However,
if it's not a grouping expression, it is not marked as
nullable leading to an assert later when the expression
is being set to null. Presence of non-grouping expressions
other than aggregates is allowed only when the full group by
mode is turned off.
We now set all columns in select list as nullable in the
presence of ROLLUP.

Change-Id: I68210d050377338eb65b385633efcb59930ea28b