Bug#38203473: Assertion `is_nullable()' in               Item_typecast_signed::val_int               at sql/item_func.cc [ITERATOR 3/3]
When there are no rows in the result set, there
is special handling to generate rollup NULLs and
super aggregates. However, if a non-grouped
expression was present in the SELECT list, these
expressions would not be set to NULL leading to
unexpected result.
We fix this by doing what is being done currently
for implicitly grouped queries on empty result set.

Change-Id: Ia6980e99cd9e591c9318618f2f37da7b31ea8eaf