Bug#38260769 ndb_restore cannot handle backup with large BackupID(>2147483647)
Backport to 7.6

Follow up fix to allow full set of unsigned 32-bit values for backupId
in ndb_restore.

When starting a backup with an explicit backup id, ndb_mgm limits to
the range 1..2^32-2, while mgmapi limits to the range 1..2^32-1.

When starting a backup without an explicit backup id, the data nodes
increment the last backup id, which is returned to the user modulo
2^32.

Therefore it is possible to reach backup ids 2^32-1 and 0, though not
by explicitly supplying them.

ndb_restore is further modified to allow backups created with
these ids to be restored.

ndb_restore_misc + ndb_restore_compat_compression are modified to
 - Show that ndb_restore can now restore filesets with backup
   ids 0 and 2^32-1.
 - Obtain only 32 bit backup ids via SYSTAB_0 / ndb_backup_id.inc
 - Cleanup backup files to avoid problems with id reuse

Change-Id: Icc581e84a93cdd1c140f0d2839e4800a5244ba5b