Bug#36714667: EstimateEqualPredicateSelectivity is called several times for same join cond [2, sargable, noclose]
Previously, the hypergraph optimizer sometimes called
EstimateSelectivity() redundantly on sargable join predicates, leading
to unnecessary computations and potential inconsistencies in row
estimation.

This happened for predicates on joins that were not part of a cycle in
the hypergraph. For join predicates that _were_ part of a cycle,
PossiblyAddSargableCondition() used the selectivity estimate stored in
JoinHypergraph::predicates. Non-cycle join predicates, however, are
not in JoinHypergraph::predicates, and their selectivities were
therefore calculated afresh with EstimateSelectivity(), even though
these selectivities had already been calculated and stored in the
RelationalExpression tree during hypergraph construction.

This patch makes PossiblyAddSargableCondition() get the selectivity
from the RelationalExpression representing the join from which the
sargable condition was pushed, and in which the join condition
selectivity has already been estimated and cached. In order to achieve
that, RelationalExpression::m_pushable_conditions and
JoinHypergraph::Node::m_pushable_conditions were changed to store a
pointer to the RelationalExpression from which the condition was
pushed. Previously, they stored which conditions to push, but not
where they were pushed from. With this extra information, the
selectivity can be read from the RelationalExpression object instead
of being recalculated.

Change-Id: I6fbcd6f11289ab914db40d64dc9e63ac42d7258e