WL#16210 Remove unused upgrade code after WL#15924
A lot of code is unreachable since we no longer support
upgrading from old versions of MySQL.

This patch removes much of this code.
Some of the functions appeared to be used in ut0test code
so this patch replaces the calls with modern equivalents.
Some were "used" in api0api.cc but actually the code was
unreachable.

Change-Id: Ie4899dff1ef88f0d1642ba300cc7ea61252d0b5b