BUG#36445818 P_S doesn't record SQL or digest for prepared statements for Non-mysql client
Problem

Prepared statements using the binary protocol,
in particular:
- COM_STMT_PREPARE
- COM_STMT_EXECUTE
- COM_STMT_CLOSE
are instrumented with the performance schema,
but the instrumentation does not capture properties like:
- the SQL text associated with the statement, when it exists
- the DIGEST and DIGEST_TEXT of the statement, when it exists

This makes troubleshooting applications difficult,
and reduces the value of the performance schema instrumentation.

Root cause

The instrumentation is incomplete,
in case of statements using the binary protocol
(i.e., when not the mysql client but a mysql connector)

Fix

In the prepared statements implementation,
in class Prepared_statement:
- add a member to hold the (possibly rewritten) query string
- add a member to hold a digest
- add members to hold the token array used to represent a digest

In the PREPARE implementation,
revisit the performance schema instrumentation:
- capture the query string for the statement
- record the query string, possibly rewritten, in the performance schema
  statement instrumentation. Note that this is done even when the prepare
  later fails, so that debugging failed prepare is possible.
- record the statement digest in the performance schema instrumentation
- keep a copy of the query string and digest in the prepared statement itself,
  to make it available during COM_STMT_EXECUTE and COM_STMT_CLOSE.

In the EXECUTE and DEALLOCATE PREPARE implementation,
revisit the performance schema instrumentation:
- retrieve the query string and digest from the prepared statement,
  and use it to instrument the statement with the performance schema.

The net effect is as follows:

For prepared statements using the binary protocol:
- COM_STMT_PREPARE
- COM_STMT_EXECUTE
- COM_STMT_CLOSE
there is now instrumentation, compared to none previously,
fixing the issue reported.

For the prepared statements using the text protocol:
- SQLCOM_PREPARE
- SQLCOM_EXECUTE
- SQLCOM_DEALLOCATE_PREPARE
the existing instrumentation has been improved,
to follow the same rules as the binary protocol,
and provides more insight for monitoring.

Considering the following statements:
- PREPARE stmt FROM "SELECT * FROM t1 WHERE id = ?";
- EXECUTE stmt;
- DEALLOCATE PREPARE stmt;

Previously, the instrumentation captured respectively digests like:
- PREPARE stmt FROM ?;
- EXECUTE stmt;
- DEALLOCATE PREPARE stmt;
which gives no insight on what is really executed.

With this change, the instrumentation digests are:
- PREPARE SELECT * FROM t1 WHERE ID = ?
- EXECUTE SELECT * FROM t1 WHERE ID = ?
- DEALLOCATE SELECT * FROM t1 WHERE ID = ?
for PREPARE, EXECUTE and DEALLOCATE PREPARE,
pointing to the statement actually used.

This change is desirable, improving the value for monitoring tools,
by giving more visibility into the application execution.

Change-Id: If115a58fa9bccd1d0507791c33c56a45f0044797