WL#17037: Exempt unknown users from connection control's delay criteria
Connection control component/plugin introduces a delay if a number of
failed login attempts for a given authID goes above threshold configured
by the administrator.

This worklog implements a way for the following legal usecase to be
exempted from introducing the delay:
 - connections that do not use MySQL protocol, just probing if the
   server is up and running (for example originating from load balancer)

This commit implements the required changes only to the connection
control component.

Change-Id: I580da934f1141351e59d63b2659893382a54ceed