WL#16621  Change Thread Pool Plugin configuration defaults.
Description
-----------
This worklog introduces automatic hardware-based configuration in Thread
Pool plugin. The goal is to enable the Thread Pool to automatically
detect the underlying server hardware (on-cloud or on-premises
deployments) and set optimal defaults to deliver better performance.

New Defaults
------------
- thread_pool_size: Changed the default from 16 to MIN ( VCPU/2, 16 )
  in the range [1, 16]

- thread_pool_max_transactions_limit: Change default from 0 to
  MIN(thread_pool_size * 32, 512 ) in the range [1, 512].
  Note: Explicitly setting it to 0 shall disable
  thread_pool_max_transactions_limit.

- thread_pool_query_threads_per_group: Change default from 1 to 2. Fixed
  for all shapes.

- thread_pool_algorithm: Change default from 0 to 1. Fixed for all
  shapes.

- thread_pool_max_unused_threads: Change default from 0 to 2. Fixed for
  all shapes.

Implementation
--------------
- The plugin now uses the system_variable_source service to query the
current value of the server/plugin variables. This also allows the
plugin to know how the variable was set (e.g., through SET PERSIST,
command-line, or in the configuration file).

- If the plugin variable is set by the user, the plugin will not override it
with the new default value. This ensures that user-defined settings take
precedence over the plugin's defaults.

- If the plugin variable is not set by the user, the plugin will apply the
new default value.

- We now define the min, max, and default values for the plugin
variables. This ensures any attempts to set the values outside these
limits results in auto-correction/clamping to the nearest valid value.
Earlier, any misconfigurations resulted in server startup without the
plugin loaded.

- The plugin will now log a warning message if the plugin variable is
auto-corrected/clamped to the nearest valid value. This is to inform
users that their configuration was adjusted to a valid value.

- The plugin will auto-correct any incompatibilities raised from the new
  defaults. This can happen when the user has explicitly set a value to
  one variable that is incompatible with the new defaults of other
  variables. In such cases, the plugin will log a warning message and
  adjust the value to a compatible one.

- The variables `thread_pool_max_transactions_limit` and
  `thread_pool_query_threads_per_group` will now be persisted in the
  static section of `mysqld-auto.cnf`. This is done to ensure that all
  inter-dependent variables are persisted in the same manner and their
  values are available when Thread Pool starts.

Change-Id: I0c73599f0449261cb5409b36094dbe9fd8c5fbed