Bug#38179925: Insert into a JDV with AUTO_INCREMENT colunn is rejected even if not part of join condition
Problem: Any insert into a JDV which did not provide a value for an
AUTO_INCRMENT column was rejected. This is stricter that necessary, as
only AUTO_INCREMENT columns which are part of a join-condition are
problematic for JDVs (because it is not currently possible to obtain the
resulting AUTO_INCREMENT value and use it to infer the value for the other
column in the join condition).

Solution: Remove restriction and rely on check which ensures that PK
columns are provided a value in JSON. This is sufficient as all join-conditions
involve a PK column.

Change-Id: I1a4baedc85defa9996e4e62ab7b848231a93e947