Bug#38194951 PB2: ndb_binlog.ndb_binlog_split_rows fails on solaris/sparc
Problem:

The ndb_binlog.ndb_binlog_split_rows and
ndb_binlog.ndb_binlog_split_rows_rpltab tests uses a 12 row table which
is split into 1 to 4 row slices.  Then the sizes of one or several
slices are checked to reasonable.  The checks +-1 row from expected row
count.  That worked for the sample table used on little endian systems
but failed on big endian systems.

Fix:

Widen the acceptable interval for the row counts to intervals with at
least 95% confidence, which turned out to be enough for test to also
pass on big endian.

For the passing test ndb_binlog.ndb_binlog_split_rows_big which uses
bigger tables with 1024-6144 rows the test intervals are changed to 3
sigma confidence intervals instead of hard coded numbers (approximating
the distribution with a normal distribution).

Change-Id: I38800f552a63488b683587e90f0637e19f03a393