Bug#36955411: Hypergraph optimizer does not use index lookups for <=> predicates [post-push fix]
The fix for this bug added an assertion to verify that
LEX::using_hypergraph_optimizer() always returns true when
FindBestQueryPlan() is called. Also, it added calls to
LEX::set_using_hypergraph_optimizer() in many of the unit test cases
in order to satisfy the new assertion.

This patch removes the explicit calls to
LEX::set_using_hypergraph_optimizer() from the unit test cases, and
instead makes the test cases inherit from HypergraphOptimizerTestBase,
which already enables the hypergraph optimizer correctly. This
approach has the additional benefit that the flag is visible during
resolving, because HypergraphOptimizerTestBase enables the hypergraph
optimizer earlier.

The point-select microbenchmark is a free function, so changing the
inheritance wasn't an option, and the hypergraph optimizer still has
to be enabled manually. It was changed to enable the hypergraph
optimizer by setting an optimizer switch before resolving instead of
calling set_using_hypergraph_optimizer() after resolving. This better
matches what HypergraphOptimizerTestBase does.

Unrelated change: Unnecessary initialization of current_thd was
removed. It is already initialized by Server_initializer::SetUp().

Change-Id: Ica8d78155440b32973431de02350525d815f1ac2